x/redduct/keeper/msg_server_cartitem_test.go
package keeper

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"

	"github.com/sap200/redduct/x/redduct/types"
)

func TestCartitemMsgServerCreate(t *testing.T) {
	srv, ctx := setupMsgServer(t)
	creator := "A"
	for i := 0; i < 5; i++ {
		resp, err := srv.CreateCartitem(ctx, &types.MsgCreateCartitem{Creator: creator})
		require.NoError(t, err)
		assert.Equal(t, i, int(resp.Id))
	}
}

func TestCartitemMsgServerUpdate(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgUpdateCartitem
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdateCartitem{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateCartitem{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateCartitem{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		tc := tc
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)
			_, err := srv.CreateCartitem(ctx, &types.MsgCreateCartitem{Creator: creator})
			require.NoError(t, err)

			_, err = srv.UpdateCartitem(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}

func TestCartitemMsgServerDelete(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgDeleteCartitem
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeleteCartitem{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeleteCartitem{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "KeyNotFound",
			request: &types.MsgDeleteCartitem{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		tc := tc
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)

			_, err := srv.CreateCartitem(ctx, &types.MsgCreateCartitem{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeleteCartitem(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
